перенес отрисовку маршрута в презентер. пока еще не открываются точки
