+++ src/main.js
@@ -1,10 +1,13 @@
 import SiteMenuView from './view/site-menu.js';
 import FilterView from './view/filter.js';
-import SortView from './view/sort.js';
-import EventsContainer from './view/trip.js';
-import EventView from './view/event.js';
-import EventEditView from './view/event-edit';
-import NoEventView from './view/no-event.js';
+// import SortView from './view/sort.js';
+// import EventsContainer from './view/events-list.js';
+// import EventView from './view/event.js';
+// import EventEditView from './view/event-edit';
+// import NoEventView from './view/no-event.js';
+
+import TripPresenter from './presenter/trip.js';
+
 import {generateEvent} from './mock/event.js';
 import {render, RenderPosition} from './utils/render.js';
 
@@ -16,54 +19,57 @@ const siteMenuElement = document.querySelector('.trip-controls__navigation');
 const siteFilterElement = document.querySelector('.trip-controls__filters');
 const tripEventsElement = document.querySelector('.trip-events');
 
-const renderEvent = (eventListElement, event) => {
-  const eventComponent = new EventView(event);
-  const eventEditComponent = new EventEditView(event);
+// const renderEvent = (eventListElement, event) => {
+//   const eventComponent = new EventView(event);
+//   const eventEditComponent = new EventEditView(event);
 
-  const openPoint = () => {
-    eventListElement.replaceChild(eventEditComponent.getElement(), eventComponent.getElement());
-  };
+//   const openPoint = () => {
+//     eventListElement.replaceChild(eventEditComponent.getElement(), eventComponent.getElement());
+//   };
 
-  const closePoint = () => {
-    eventListElement.replaceChild(eventComponent.getElement(), eventEditComponent.getElement());
-  };
+//   const closePoint = () => {
+//     eventListElement.replaceChild(eventComponent.getElement(), eventEditComponent.getElement());
+//   };
 
-  const onEscKeyDown = (evt) => {
-    if (evt.key === 'Escape' || evt.key === 'Esc') {
-      evt.preventDefault();
-      closePoint();
-      document.removeEventListener('keydown', onEscKeyDown);
-    }
-  };
+//   const onEscKeyDown = (evt) => {
+//     if (evt.key === 'Escape' || evt.key === 'Esc') {
+//       evt.preventDefault();
+//       closePoint();
+//       document.removeEventListener('keydown', onEscKeyDown);
+//     }
+//   };
 
-  eventComponent.setOpenButtonClickHandler(() => {
-    openPoint();
-    document.addEventListener('keydown', onEscKeyDown);
-  });
+//   eventComponent.setOpenButtonClickHandler(() => {
+//     openPoint();
+//     document.addEventListener('keydown', onEscKeyDown);
+//   });
 
-  eventEditComponent.setCloseButtonClickHandler(() => {
-    closePoint();
-    document.removeEventListener('keydown', onEscKeyDown);
-  });
+//   eventEditComponent.setCloseButtonClickHandler(() => {
+//     closePoint();
+//     document.removeEventListener('keydown', onEscKeyDown);
+//   });
 
-  eventEditComponent.setFormSubmitHandler(() => {
-    closePoint();
-    document.removeEventListener('keydown', onEscKeyDown);
-  });
+//   eventEditComponent.setFormSubmitHandler(() => {
+//     closePoint();
+//     document.removeEventListener('keydown', onEscKeyDown);
+//   });
 
-  render(eventListElement, eventComponent.getElement(), RenderPosition.BEFOREEND);
-};
+//   render(eventListElement, eventComponent.getElement(), RenderPosition.BEFOREEND);
+// };
 
 render(siteMenuElement, new SiteMenuView().getElement(), RenderPosition.BEFOREEND);
 render(siteFilterElement, new FilterView().getElement(), RenderPosition.BEFOREEND);
 
-if (events.length > 0) {
-  render(tripEventsElement, new SortView().getElement(), RenderPosition.BEFOREEND);
-  render(tripEventsElement, new EventsContainer().getElement(), RenderPosition.BEFOREEND);
-  const eventsList = document.querySelector('.trip-events__list');
-  for (let i = 0; i < EVENT_COUNT; i++) {
-    renderEvent(eventsList, events[i]);
-  }
-} else {
-  render(tripEventsElement, new NoEventView().getElement(), RenderPosition.BEFOREEND);
-}
+// if (events.length > 0) {
+//   render(tripEventsElement, new SortView().getElement(), RenderPosition.BEFOREEND);
+//   render(tripEventsElement, new EventsContainer().getElement(), RenderPosition.BEFOREEND);
+//   const eventsList = document.querySelector('.trip-events__list');
+//   for (let i = 0; i < EVENT_COUNT; i++) {
+//     renderEvent(eventsList, events[i]);
+//   }
+// } else {
+//   render(tripEventsElement, new NoEventView().getElement(), RenderPosition.BEFOREEND);
+// }
+
+const tripPresenter = new TripPresenter(tripEventsElement);
+tripPresenter.init(events);
